requirements.txt
gradio==3.28.3
gradio_client==0.1.4
Pillow==9.5.0
pymilvus==2.2.15
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0